fix methods for deleting permissions

+++ src/main/java/com/softserve/rms/service/implementation/PermissionManagerServiceImpl.java
@@ -3,11 +3,13 @@ package com.softserve.rms.service.implementation;
 import com.softserve.rms.constants.ErrorMessage;
 import com.softserve.rms.dto.PermissionDto;
 import com.softserve.rms.entities.ResourceTemplate;
-import com.softserve.rms.exceptions.security.DeniedAccessException;
+import com.softserve.rms.exceptions.PermissionException;
+import com.softserve.rms.security.mappers.PermissionMapper;
 import com.softserve.rms.service.PermissionManagerService;
-import com.softserve.rms.util.PermissionMapper;
+import com.softserve.rms.util.Formatter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.acls.domain.BasePermission;
 import org.springframework.security.acls.domain.GrantedAuthoritySid;
 import org.springframework.security.acls.domain.ObjectIdentityImpl;
@@ -16,18 +18,22 @@ import org.springframework.security.acls.model.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.security.Principal;
 import java.util.List;
 
 @Service
+@PreAuthorize("hasRole('MANAGER')")
 public class PermissionManagerServiceImpl implements PermissionManagerService {
 
     private final MutableAclService mutableAclService;
     private final PermissionMapper permissionMapper;
+    private final Formatter formatter;
 
     @Autowired
-    public PermissionManagerServiceImpl(MutableAclService mutableAclService, PermissionMapper permissionMapper) {
+    public PermissionManagerServiceImpl(MutableAclService mutableAclService, PermissionMapper permissionMapper, Formatter formatter) {
         this.mutableAclService = mutableAclService;
         this.permissionMapper = permissionMapper;
+        this.formatter = formatter;
     }
 
     @Transactional
@@ -79,15 +85,26 @@ public class PermissionManagerServiceImpl implements PermissionManagerService {
         }
     }
 
+    /**
+     * Method deletes one permission for
+     * certain user.
+     *
+     * @param permissionDto transfer information about certain ACE
+     * @param principal authenticated user
+     */
     @Transactional
     @Override
-    public void closePermissionForCertainUser(PermissionDto permissionDto) {
+    public void closePermissionForCertainUser(PermissionDto permissionDto, Principal principal) {
         MutableAcl acl;
-        PrincipalSid sid = new PrincipalSid(permissionDto.getRecipient());
+        Sid sid = permissionDto.isPrincipal()
+                ? new PrincipalSid(permissionDto.getRecipient()) : new GrantedAuthoritySid(permissionDto.getRecipient());
         ObjectIdentity objectIdentity = new ObjectIdentityImpl(ResourceTemplate.class, permissionDto.getResTempId());
         int acePosition = 0;
         try {
             acl = (MutableAcl) mutableAclService.readAclById(objectIdentity);
+            if(!formatter.sidFormatter(acl.getOwner().toString()).equals(principal.getName())) {
+                throw new PermissionException(ErrorMessage.DENIED_ACCESS.getMessage());
+            }
             List<AccessControlEntry> aces = acl.getEntries();
             for (AccessControlEntry entry : aces) {
                 if (entry.getSid().equals(sid)
@@ -98,19 +115,30 @@ public class PermissionManagerServiceImpl implements PermissionManagerService {
                 acePosition++;
             }
             mutableAclService.updateAcl(acl);
-        } catch (NotFoundException | DataAccessException e) {
-            throw new DeniedAccessException(ErrorMessage.DENIED_ACCESS.getMessage());
+        } catch (NotFoundException e) {
+            throw new PermissionException(ErrorMessage.PERMISSION_NOT_FOUND.getMessage());
         }
     }
 
+    /**
+     * Method deletes ACL for certain resource template.
+     *
+     * @param resourceId id of resource template
+     * @param principal authenticated user
+     */
     @Transactional
     @Override
-    public void closeAllPermissionsToResource(Long resourceId) {
-        ObjectIdentity objectIdentity = new ObjectIdentityImpl(resourceId);
+    public void closeAllPermissionsToResource(Long resourceId, Principal principal) {
+        MutableAcl acl;
         try {
+            ObjectIdentity objectIdentity = new ObjectIdentityImpl(ResourceTemplate.class, resourceId);
+            acl = (MutableAcl) mutableAclService.readAclById(objectIdentity);
+            if(!formatter.sidFormatter(acl.getOwner().toString()).equals(principal.getName())) {
+                throw new PermissionException(ErrorMessage.DENIED_ACCESS.getMessage());
+            }
             mutableAclService.deleteAcl(objectIdentity, false);
-        } catch (NotFoundException | DataAccessException e) {
-            throw new DeniedAccessException(ErrorMessage.DENIED_ACCESS.getMessage());
+        } catch (NotFoundException e) {
+            throw new PermissionException(ErrorMessage.PERMISSION_NOT_FOUND.getMessage());
         }
     }
 }
